Moved server logic to server file isolated file
