+++ .gitignore
@@ -6,3 +6,4 @@ src/__pycache__
 /src/server/__pycache__
 /src/client/__pycache__
 /.vscode
+/__pycache__
